requirements.txt
absl-py==0.7.0
asgiref==3.2.7
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==3.0.7
django-extensions==3.0.5
entrypoints==0.3
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
iamport-rest-client==0.8.2
idna==2.10
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.2.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.0
opencv-python==4.1.2.30
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
Pygments==2.3.1
pyparsing==2.4.7
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2020.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.24.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.1
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.10
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
